Move theme in internal layouts
