+++ assets/js/anatole-theme-switcher.js
