+++ assets/js/jquery-migrate.js
@@ -8,14 +8,14 @@
     if (typeof define === "function" && define.amd) {
 
         // AMD. Register as an anonymous module.
-        define(["jquery"], function (jQuery) {
+        define(["assets/js/jquery"], function (jQuery) {
             return factory(jQuery, window);
         });
     } else if (typeof module === "object" && module.exports) {
 
         // Node/CommonJS
         // eslint-disable-next-line no-undef
-        module.exports = factory(require("jquery"), window);
+        module.exports = factory(require("assets/js/jquery"), window);
     } else {
 
         // Browser globals
