+++ layouts/_default/single.html
@@ -63,20 +63,6 @@
         </div>
     </div>
 
-    {{ if eq .Type "post"}}
-    {{- if .Site.DisqusShortname -}}
-    <div id="fb_comments_container">
-        <h2>{{ i18n "comments" }}</h2>
-        {{ template "_internal/disqus.html" . }}
-    </div>
-    {{- end -}}
-    {{- if .Site.Params.utterances.repo -}}
-    <div id="fb_comments_container">
-        <h2>{{ i18n "comments" }}</h2>
-        {{ partial "comments/utterances.html" . }}
-    </div>
-    {{- end -}}
-    {{ end }}
 </div>
 
 {{ end }}
