+++ layouts/partials/sidebar.html
@@ -2,7 +2,7 @@
     <div class="logo-title">
         <div class="title">
             <img src="{{ .Site.Params.profilePicture | absURL }}" alt="profile picture">
-            <h3 title=""><a href="/">{{ .Site.Params.Title }}</a></h3>
+            <h3 title=""><a href="/static">{{ .Site.Params.Title }}</a></h3>
             <div class="description">
                 <p>{{ .Site.Params.description }}</p>
             </div>
