--- themes/anatole/archetypes/post.md
@@ -1,8 +1,0 @@
----
-title: "{{ replace .Name "-" " " | title }}"
-date: {{ .Date }}
-Description: ""
-Tags: []
-Categories: []
-DisableComments: false
----
