--- themes/anatole/exampleSite/content/_index.md
@@ -1,24 +1,0 @@
-+++
-author = "Hugo Authors"
-+++
-
-<!--
-This file is left intentionally empty by default to be backward compatible with initial theme setup.
-
-Although the theme has advanced a little bit and it now allows to specify the content on the main page (even if the list of posts/articles is not intended).
-This can be:
-- with the list of posts/articles (default: `mainSections = ["post"]) or
-- without the list of posts/articles (by setting `mainSections = [""]`)
-
-Markdown supported, ie:
-
-```
-# Welcome
-
-- Hugo :rocket:
-- Hugo theme :rocket:
-
-Don't forget to check the README.md file!
-```
-
--->
