--- themes/anatole/exampleSite/content/archives.md
@@ -1,5 +1,0 @@
----
-date: 2019-05-28
-type: section
-layout: "archives"
----
