--- themes/anatole/exampleSite/content/contact.md
@@ -1,7 +1,0 @@
----
-author: Hugo Authors
-title: Contact
-date: 2019-03-08
-description: Contact Page
-contact: true
----
