--- themes/anatole/exampleSite/content/post/_index.md
@@ -1,6 +1,0 @@
-+++
-aliases = ["posts", "articles", "blog", "showcase", "docs"]
-title = "Posts"
-author = "Hugo Authors"
-tags = ["index"]
-+++
