--- themes/anatole/i18n/en.toml
@@ -1,24 +1,0 @@
-[category]
-other = "category"
-
-[tag]
-other = "tag"
-
-[reading_time]
-one = "One-minute read"
-other = "{{ .Count }}-minute read"
-
-[page_not_found]
-other = "Page Not Found"
-
-[page_does_not_exist]
-other = "Sorry, this page does not exist."
-
-[head_back]
-other = "You can head back to <a href=\"{{ . }}\">homepage</a>."
-
-[comments]
-other = "comments"
-
-[send]
-other = "Send"
