--- themes/anatole/i18n/fi.toml
@@ -1,24 +1,0 @@
-[category]
-other = "kategoria"
-
-[tag]
-other = "merkki"
-
-[reading_time]
-one = "Yksi lukuminuutti"
-other = "{{ .Count }} lukuminuuttia"
-
-[page_not_found]
-other = "Sivua ei löydetty"
-
-[page_does_not_exist]
-other = "Valitettavasti tätä sivua ei ole olemassa."
-
-[head_back]
-other = "Voit palata takaisin <a href=\"{{ . }}\">kotisivulle</a>."
-
-[comments]
-other = "kommentit"
-
-[send]
-other = "Lähetä"
