--- themes/anatole/i18n/fr.toml
@@ -1,24 +1,0 @@
-[category]
-other = "catégorie"
-
-[tag]
-other = "tag"
-
-[reading_time]
-one = "Une minute de lecture"
-other = "{{ .Count }} minutes de lecture"
-
-[page_not_found]
-other = "Page Non Trouvée"
-
-[page_does_not_exist]
-other = "Désolé, cette page n'existe pas."
-
-[head_back]
-other = "Vous pouvez revenir à <a href=\"{{ . }}\">l'accueil</a>."
-
-[comments]
-other = "commentaire"
-
-[send]
-other = "Envoyer"
