--- themes/anatole/i18n/it.toml
@@ -1,24 +1,0 @@
-[category]
-other = "categoria"
-
-[tag]
-other = "tag"
-
-[reading_time]
-one = "Tempo di lettura: Un minuto"
-other = "Tempo di lettura: {{ .Count }} Minuti"
-
-[page_not_found]
-other = "Pagina non trovata"
-
-[page_does_not_exist]
-other = "Spiacente, questa pagina non esiste."
-
-[head_back]
-other = "Puoi tornare alla <a href=\"{{ . }}\">homepage</a>."
-
-[comments]
-other = "commenti"
-
-[send]
-other = "Spedire"
