--- themes/anatole/i18n/zh-cn.toml
@@ -1,24 +1,0 @@
-[category]
-other = "分类"
-
-[tag]
-other = "标签"
-
-[reading_time]
-one = "阅读时间 1 分钟"
-other = "阅读时间 {{ .Count }} 分钟"
-
-[page_not_found]
-other = "页面未找到"
-
-[page_does_not_exist]
-other = "抱歉，此页面不存在"
-
-[head_back]
-other = "返回<a href=\"{{ . }}\">首页</a>."
-
-[comments]
-other = "评论"
-
-[send]
-other = "发送"
