--- themes/anatole/i18n/zh-tw.toml
@@ -1,24 +1,0 @@
-[category]
-other = "分類"
-
-[tag]
-other = "標籤"
-
-[reading_time]
-one = "閱讀時間 1 分鐘"
-other = "閱讀時間 {{ .Count }} 分鐘"
-
-[page_not_found]
-other = "找不到頁面"
-
-[page_does_not_exist]
-other = "此頁面不存在"
-
-[head_back]
-other = "返回 <a href=\"{{ . }}\">首頁</a>."
-
-[comments]
-other = "註解"
-
-[send]
-other = "發送"
