--- themes/anatole/layouts/partials/comments/utterances.html
@@ -1,10 +1,0 @@
-<script src="https://utteranc.es/client.js" 
-        repo="{{ .Site.Params.utterances.repo }}"
-        issue-term="{{ .Site.Params.utterances.issueTerm }}"
-        theme="{{ .Site.Params.utterances.theme }}" 
-        {{ with .Site.Params.utterances.label }}
-        label="{{ . }}"
-        {{ end }}
-        crossorigin="anonymous" 
-        async>
-</script>
