--- themes/anatole/theme.toml
@@ -1,21 +1,0 @@
-# theme.toml template for a Hugo theme
-# See https://github.com/gohugoio/hugoThemes#themetoml for an example
-
-name = "Anatole"
-license = "MIT"
-licenselink = "https://github.com/lxndrblz/anatole/blob/master/LICENSE"
-description = "Anatole is a beautiful minimalist two-column hugo theme based on farbox-theme-Anatole."
-homepage = "https://github.com/lxndrblz/anatole/"
-tags = ["blog", "responsive", "clean", "minimalist", "minimal", "personal", "dark"]
-features = ["blog", "seo", "responsive", "mobile", "disqus", "fontawesome", "analytics", "math"]
-min_version = "0.53.0"
-
-[author]
-  name = "Alexander Bilz"
-  homepage = "https://alexbilz.com"
-
-# If porting an existing theme
-[original]
-  author = "Cai Cai"
-  homepage = "https://www.caicai.me/"
-  repo = "https://github.com/hi-caicai/farbox-theme-Anatole"
