--- themes/plausible-hugo/i18n/en.toml
@@ -1,2 +1,0 @@
-[plausible_gitstar]
-other = "*** Please, consider leaving a star on Github if you like this plausible-hugo module.\n Remove this warning by adding 'gitstar = false' in your plausible config"
