--- themes/plausible-hugo/i18n/fr.toml
@@ -1,3 +1,0 @@
-
-[plausible_gitstar]
-other = "*** Merci de liker (star) ce module sur Github si vous appréciez plausible-hugo.\n Supprimez ce warning en ajoutant 'gitstar = false' dans votre config plausible"
